refactor(sdk-derive): depricate solidity_storage macro
